BUG : minimal fix for date x-values in stem plots
closes #2602

+++ lib/matplotlib/axes/_axes.py
@@ -2263,7 +2263,7 @@ class Axes(_AxesBase):
         self.hold(True)
 
         # Assume there's at least one data array
-        y = np.asarray(args[0], dtype=np.float)
+        y = np.asarray(args[0])
         args = args[1:]
 
         # Try a second one
